Remove all my hard work and add timestamp before db insert
