+++ link.go
@@ -1,24 +1,36 @@
 package main
 
 import (
+	"errors"
+	"net/http"
+	"net/url"
 	"time"
 )
 
 type Link struct {
-	ID        string `json:"id" form:"id"`
-	URL       string `json:"url" form:"url,omitempty"`
-	Timestamp time.Time
+	ID        string    `json:"id" form:"id"`
+	URL       string    `json:"url" form:"url,omitempty"`
+	Timestamp time.Time `json:"@timestamp" form:"@timestamp"`
 }
 
 func (l *Link) String() string {
 	return l.URL
 }
 
-// NewLink creates a new link with a timestamp of the current time
-func NewLink(URL string, ID string) *Link {
-	return &Link{
-		URL:       URL,
-		ID:        ID,
-		Timestamp: time.Now(),
+func (link *Link) Render(w http.ResponseWriter, r *http.Request) error {
+	return nil
+}
+
+func (link *Link) Bind(r *http.Request) error {
+	if link.URL == "" {
+		return errors.New("Malformed URL")
 	}
+	url, err := url.Parse(link.URL)
+	if err != nil {
+		return err
+	}
+	if url.Host == "" || url.Scheme == "" {
+		return errors.New("Malformed URL")
+	}
+	return nil
 }
